Avoid broadcasting a second stop event
Tweak logic handling a stop event from the navigation API to avoid
the accidental sending of another stop event. This avoids clearing
a subsequent waypoint set that may be in flight, which was seen in
practice when doing a stop and a waypoint set from the POI
application.

Bug-AGL: SPEC-3079

Signed-off-by: Scott Murray <scott.murray@konsulko.com>
Change-Id: I12d93a15d43591d8841ea6a312e7d017f136c86d