Avoid broadcasting a second stop event
Tweak logic handling a stop event from the navigation API to avoid
the accidental sending of another stop event. This avoids clearing
a subsequent waypoint set that may be in flight, which was seen in
practice when doing a stop and a waypoint set from the POI
application.

Bug-AGL: SPEC-3079

Signed-off-by: Scott Murray <scott.murray@konsulko.com>
Change-Id: Ic738e83fae1d312377f05aec4888d84572a44f4b