Update status signal states for VSS 3.1.1
The states for Vehicle.Cabin.Infotainment.Navigation.State are now
uppercase with the update to VSS 3.1.1, update calls into the
Navigation API from libqtappfw to match.  Also comment out use of
the currently unsupported start/stop states in the status event
signal handler.  Those were removed with the switch to using VSS,
and if they are needed again down the road they will need to be
implemented somewhat differently, document that in a comment.

Bug-AGL: SPEC-4761

Signed-off-by: Scott Murray <scott.murray@konsulko.com>
Change-Id: I02863bb3754ba984dda7c9c28c77cf0f184f92cd